fix [rewriting arc insertion]: new arc GUI
The arc definition is not anymore "predefined"
but the GUI is initiated when defining the arc.

Intraface has an actual node inserted, as before,
but the transfer variant is identical on both sides.
So for example Fick diffusion
GUI only shows possible variants.

some small fixes in the main window:
  radio button for controlling editing mode
  has been adjuste.
  Kind of works except initialisation.
  We should probably define a default or status
  later can then be initialised.

Missing: variant GUI for interfaces.