fix [equation editor]: interface equation definition
interface definitions are done on the top level of inter-branches
in the ontology.

Thus the picking table must show not just the selected interface
but all variables up and DOWN the branch.
Up is built into the design of the ontology, as variables are inherited
But down had to be included.