fix(bl): Fix a bug related to the index structures
Some variables were not showing in the add variables phase of the entity editing in the Behaviour Linker. This was due a check of the size of the index structures field in the variable class. After switching to multilinear the index structures list got bigger in the majority of cases and that was not accounted for.