fix(matlab): Fix the index ordering
A new parameter - indexOrder - is added to the constructor of the MultiDimVar class, a cell array containing the ordered index labels. Whenever a new instance of the class is created the indexLabels are automatically ordered using the indexOrder as reference. If there is a rearrangement of the indexLabels the value gets permuted to keep consistency.