feat [equation editor]: @ reduce product operator
The need for a reduced product operator for which
one defines the index over which is being reduced
came about in connection with interfaces.

Mathematically this is a tensor operation for
which one would get two idential index sets,
which is not allowed, so only one, namely the one
indicated index set is reduced.
Corresponds to Einstein indexing.