new [equation editor/variable]: interface remember source
The automatic generation needs improvement
The interface-related equations now "remember" the source
If data are transferred through an interface,
the variable will remember if it came from a node or an arc.
It will also remember if it came from an interface.

fix-variable ontology etc was updated

check on imported variables added