fix [RDF_Graph] : variable identifiers
Issue was that the name of the variable was in the IRI definition. It was changed to the variable ID, and the name is added to the variable as a separate entry.

Also the creation and modification date has been added, allowing for checking if things have changed. Currently not used.